mcp-servers/bio-sequence-mcp/pyproject.toml
[project]
name = "bio-sequence-mcp"
version = "0.1.0"
description = "Bio Sequence Analysis MCP Server - 提供DNA/RNA/蛋白质序列分析功能"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=0.9.0",
    "biopython>=1.83",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
bio-sequence-mcp = "sequence_server:main"
